moved inner classes into their own file
for https://github.com/FieldDB/FieldDB/issues/2085
